requirements.txt
yagmail
pathlib
arrow
jinja2
confuse
pymongo
python-dotenv
beautifulsoup4
pandas